requirements.txt
# Core ML and Deep Learning
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
scipy>=1.7.0

# Computer Vision and Image Processing
Pillow>=8.0.0
scikit-image>=0.18.0
opencv-python>=4.5.0

# Data Processing and Analysis
pandas>=1.3.0
scikit-learn>=1.0.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Utilities
tqdm>=4.62.0
PyYAML>=6.0
click>=8.0.0

# Development and Testing
pytest>=6.0.0
black>=22.0.0
flake8>=4.0.0

# Jupyter (optional, for notebooks)
jupyter>=1.0.0
ipykernel>=6.0.0
